refactor: header layout and design
