+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import "./App.css";
 import Editor from "../Content/Editor";
-import Header from "../Header/Header";
+import Header from "../Header/";
 import Menu from "../Menu/Menu";
 import Preview from "../Content/Preview";
 import Sidebar from "../Sidebar";
