--- src/components/Header/Header.css
@@ -1,16 +1,0 @@
-.layout-header {
-  border-bottom: 1px solid var(--color-border);
-  padding: 0 1rem;
-}
-
-/* Main Component */
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-/* Header Icon */
-.header-icon {
-  margin-left: 1.5rem;
-}
