--- src/components/Header/Header.js
@@ -1,36 +1,0 @@
-import React from 'react';
-
-const Header = ({ togglePreview, deleteNote, isPreviewDisplayed }) => {
-    return (
-        <header className="layout-header header">
-            <button className="btn btn-icon btn--left" onClick={toggleSidebar} aria-label="Toggle Sidebar">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect><line x1="9" y1="3" x2="9" y2="21"></line></svg>
-            </button>
-            <div>
-                <button className="btn btn-icon btn--right" onClick={togglePreview} aria-label="Toggle Preview">
-                    {
-                        isPreviewDisplayed ?
-                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z"></path><circle cx="12" cy="12" r="3"></circle></svg>
-                            :
-                            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19m-6.72-1.07a3 3 0 1 1-4.24-4.24"></path><line x1="1" y1="1" x2="23" y2="23"></line></svg>
-                    }
-                </button>
-                <button className="btn btn-icon btn--right header-icon" onClick={deleteNote} aria-label="Delete">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><polyline points="3 6 5 6 21 6"></polyline><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path></svg>
-                </button>
-            </div>
-        </header>
-    )
-}
-
-function toggleSidebar() {
-    const SIDEBAR = document.querySelector(".layout-sidebar");
-
-    if (SIDEBAR.dataset.display === "is-active") {
-        SIDEBAR.dataset.display = "is-hidden";
-    } else {
-        SIDEBAR.dataset.display = "is-active";
-    }
-}
-
-export default Header;
